Restructure buttons into sub-folders
