+++ src/common/buttons/deleteButton/DeleteButton.js
