+++ src/common/buttons/deleteButton/DeleteButton.stories.js
